Add html-proofer
Add in a step to the Github Actions which calls `html-proofer` to
enumerate broken links.

For now, this step will not cause the build to break (note the || true).
Instead, this is for diagnostic purposes.

In order to use this on a go forward basis, all the internal links would
need to be changed to look like Liquid links, e.g.,

https://lost-stats.github.io/Time_Series/MA-models.html

should become

{{ "/Time_Series/MA-models.html" | relative_url }}