Build automated testing pipeline
Add a py.test runner that searches for all the code blocks in the
repository and then attempts to run the python and r ones in an
appropriate docker image. Fail the test if the code block doesn't
run.

These are attached to docker images currently housed in my personal
repository space while I work out how to get a machine user setup
to build these.

Note that these tests currently take a _long_ time to run and a _lot_
of them fail. The reasons for failure are mostly:
  * An uninstalled dependency in the dockerfile
  * A code block that is separated by raw markdown.

In the latter case, I've added a feature where you can tell the
parser that two code blocks should be run sequentially in the same
instance by adding a ?example=id tag to the language name. This
format is necessary due to the fact that we use Rouge as our
lexer/syntax highlighter.

An example of this can be seen in the
Presentation/Figures/line_graphs.md
file under the Python section.