Update Scatterplots.md
Added a small bit about doing the same seaborn scatterplot in a seaborn.object plot with some added things like renaming the plot, and setting the theme to all white to get started. Also added a linear fit line so users can play with it and see the art of the possible with it! Trying to figure out how to add the photo now to this locked repo using
plot.save('seaborn_objects_scatter.png', bbox_inches='tight')