Seaborn.objects scatterplot visual upload
Figured out how. Was tired last night and was trying to upload directly on main instead of the fork I thought I was working on. :) 

I'm not foreign to github, but definitely not as well-versed as yourself, or as much as I would really like to be for my future.